+++ CHANGELOG.md
@@ -4,3 +4,43 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [1.0.0] - 2020-12-30
+
+### Description
+
+- Fully working api endpoint to validate a reference number.
+- Fully working api endpoint to get check digits from a reference.
+- Unit tests for each of these helper functions. 
+
+## [0.1.2] - 2020-12-30
+
+### Added
+
+- Fully working api endpoint to validate a reference number.
+- Unit tests for the Helper function to validate a reference number.
+
+## [0.1.1] - 2020-12-29
+
+### Fixed
+
+- Fixed a bug that when the number was too big the check digits would be incorrect.
+
+## [0.1.0] - 2020-12-29
+
+### Added
+
+- Fully working api endpoint to get check digits from a reference.
+
+## [0.0.2] - 2020-12-29
+
+### Added
+
+- Unit tests for the Helper function to get check digits from a reference.
+
+## [0.0.1] - 2020-12-28
+
+### Added
+
+- API endpoint to get check digits from a reference.
+- Helper function to get check digits from a reference.